Add generic "references file" relation type
The relation type is specific to TYPO3 files and covers 1) any file existing only in TYPO3, and 2) any file uploaded to CMIS via TYPO3's CMIS-FAL driver. It is placed in the general aspect as a way to provide default support for file relations for any record type without explicitly defining a uniquely named relation for each record type.